Update Waitable API.
Instead of passing the event mask at registratrion time, passed the mask as
part of the waiter. This makes the makes always immutable and simplifies the
architecture of waiters. This is also necessary for a future fix that will
allow the fdnotifier to keep persistent entries, as opposed to requiring
constant updates.

This change is intended to be a no-op in terms of function. The only exception
is signalfd, where this mask was abused. To handle this case, the operation of
signalfd changed to allow one layer of indirection.

PiperOrigin-RevId: 407125865